platform/mulle: Added Mulle platform.
This is the first commit of a port for Eistec's new Mulle platform.

The new Mulle is based on a Freescale Kinetis K60 (ARM Cortex M4) and
has some on board peripherals:

- Atmel AT86RF212 (rf230) radio
- FM25L04B 4 Kbit FRAM
- M25P16 16 MBit flash memory
- LIS3DH accelerometer
- LiPo battery charging circuit

Maintainer for the new port is Joakim Gebart <joakim.gebart@eistec.se>

See http://www.eistec.se or info@eistec.se for more information about
hardware and software.