cpu/arm/k60: Add check for CPUID.
Added an assert to verify that the CPU revision built for is the same
CPU revision that we are running on in core_clocks_init()

If we try to set up the clocks using the wrong version of the code
we will end up in an infinite loop waiting for the MCG status
registers to set the correct bits.

Failing the check will issue a breakpoint instruction (called SIGTRAP
in gdb) and then get stuck in an infinite loop.

Signed-off-by: Joakim Gebart <joakim.gebart@eistec.se>